requirements.txt
Flask==2.3.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
Werkzeug==2.3.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.3
python-dotenv==1.0.0
graphene==3.3
flask-cors==4.0.0
lxml==4.9.3
